# openapi/boards.yaml
openapi: '3.0.3'
info:
  title: Boards API
  version: '1.0.0'
components:
  parameters:
    boardId:
      name: boardId
      description: The ID of the board
      in: path
      required: true
      schema:
        $ref: './common.yaml#/components/schemas/MongoObjectId'
...
  schemas:
    BoardLayout:
      type: string
      description: The layout type of the board
      enum:
        - grid
        - list
    KeyResponse:
      title: KeyResponse
      type: object
      properties:
        id:
          $ref: './common.yaml#/components/schemas/MongoObjectId'
        boardId:
          $ref: './common.yaml#/components/schemas/MongoObjectId'
        name:
          $ref: './common.yaml#/components/schemas/DefaultName'
        description:
          $ref: './common.yaml#/components/schemas/DefaultDescription'
        audioMediaId:
          $ref: './common.yaml#/components/schemas/MongoObjectId'
        audioUrl:
          type: string
          format: uri
          example: https://example.com/audio.mp3
        imageMediaId:
          $ref: './common.yaml#/components/schemas/MongoObjectId'
        imageUrl:
          type: string
          format: uri
          example: https://example.com/image.png
        hotKey:
          type: string
          minLength: 1
          maxLength: 1
          example: D
        createdAt:
          $ref: './common.yaml#/components/schemas/Timestamp'
        updatedAt:
          $ref: './common.yaml#/components/schemas/Timestamp'
    BoardListResponse:
      title: GetAllBoardsResponse
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          minLength: 1
          maxLength: 255
          example: Boards retrieved successfully
        data:
          type: object
          properties:
            boards:
              type: array
              items:
                $ref: '#/components/schemas/BoardResponse'
            pagination:
              $ref: './common.yaml#/components/schemas/PaginationMeta'
    BoardResponse:
      title: BoardResponse
      type: object
      properties:
        id:
          $ref: './common.yaml#/components/schemas/MongoObjectId'
        authorId:
          $ref: './common.yaml#/components/schemas/MongoObjectId'
        name:
          $ref: './common.yaml#/components/schemas/DefaultName'
        description:
          $ref: './common.yaml#/components/schemas/DefaultDescription'
        layout:
          $ref: '#/components/schemas/BoardLayout'
        imageUrl:
          $ref: './common.yaml#/components/schemas/ImageUrl'
        collaborators:
          type: array
          items:
            $ref: '#/components/schemas/CollaboratorResponse'
        keys:
          type: array
          items:
            $ref: '#/components/schemas/KeyResponse'
        createdAt:
          $ref: './common.yaml#/components/schemas/Timestamp'
        updatedAt:
          $ref: './common.yaml#/components/schemas/Timestamp'
    CreateBoardRequest:
      type: object
      properties:
        name:
          $ref: './common.yaml#/components/schemas/DefaultName'
        description:
          $ref: './common.yaml#/components/schemas/DefaultDescription'
        layout:
          $ref: '#/components/schemas/BoardLayout'
        imageUrl:
          $ref: './common.yaml#/components/schemas/ImageUrl'
        authorId:
          $ref: './common.yaml#/components/schemas/MongoObjectId'
    CreateBoardResponse:
      title: CreateBoardResponse
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          minLength: 1
          maxLength: 255
          example: Board created successfully
        data:
          $ref: '#/components/schemas/BoardResponse'
    UpdateBoardRequest:
      type: object
      properties:
        name:
          $ref: './common.yaml#/components/schemas/DefaultName'
        description:
          $ref: './common.yaml#/components/schemas/DefaultDescription'
        layout:
          $ref: '#/components/schemas/BoardLayout'
        imageUrl:
          $ref: './common.yaml#/components/schemas/ImageUrl'
    UpdateBoardResponse:
      title: UpdateBoardResponse
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          minLength: 1
          maxLength: 255
          example: Board updated successfully
        data:
          $ref: '#/components/schemas/BoardResponse'
    DeleteBoardResponse:
      title: DeleteBoardResponse
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          minLength: 1
          maxLength: 255
          example: Board deleted successfully
    CollaboratorRole:
      type: string
      enum:
        - owner
        - editor
        - viewer
    CollaboratorResponse:
      title: CollaboratorResponse
      type: object
      properties:
        id:
          $ref: './common.yaml#/components/schemas/MongoObjectId'
        userId:
          $ref: './common.yaml#/components/schemas/MongoObjectId'
        email:
          $ref: './common.yaml#/components/schemas/UserEmail'
        name:
          $ref: './common.yaml#/components/schemas/DefaultName'
        role:
          $ref: '#/components/schemas/CollaboratorRole'
        imageUrl:
          type: string
          description: User's profile image URL
          example: "https://example.com/avatar.jpg"
        createdAt:
          $ref: './common.yaml#/components/schemas/Timestamp'
        updatedAt:
          $ref: './common.yaml#/components/schemas/Timestamp'
...
paths:
  /boards:
    get:
      summary: Get all boards
      description: Returns all boards for a user
      operationId: getAllBoards
      tags: [boards]
      parameters:
        - $ref: './common.yaml#/components/parameters/page'
        - $ref: './common.yaml#/components/parameters/limit'
      responses:
        '200':
          description: The boards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                title: GetAllBoardsUnauthorizedResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                title: GetAllBoardsErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                title: GetAllBoardsServerErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
    post:
      summary: Create a board
      description: Creates a new board for a user
      operationId: createBoard
      tags: [boards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardRequest'
      responses:
        '201':
          description: The board created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBoardResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                title: CreateBoardUnauthorizedResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                title: CreateBoardErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                title: CreateBoardServerErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
  /boards/{boardId}:
    get:
      summary: Get a board by ID
      description: Returns the board with the given ID
      operationId: getBoardById
      tags: [boards]
      parameters:
        - $ref: '#/components/parameters/boardId'
      responses:
        '200':
          description: The board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                title: GetBoardByIdUnauthorizedResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '404':
          description: Board not found
          content:
            application/json:
              schema:
                title: GetBoardByIdNotFoundResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                title: GetBoardByIdErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                title: GetBoardByIdServerErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
    patch:
      summary: Update a board
      description: Updates the board with the given ID
      operationId: updateBoard
      tags: [boards]
      parameters:
        - $ref: '#/components/parameters/boardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBoardRequest'
      responses:
        '200':
          description: The board updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateBoardResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                title: UpdateBoardUnauthorizedResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '404':
          description: Board not found
          content:
            application/json:
              schema:
                title: UpdateBoardNotFoundResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                title: UpdateBoardErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                title: UpdateBoardServerErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a board
      description: Deletes the board with the given ID
      operationId: deleteBoard
      tags: [boards]
      parameters:
        - $ref: '#/components/parameters/boardId'
      responses:
        '200':
          description: The board deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteBoardResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                title: DeleteBoardUnauthorizedResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '404':
          description: Board not found
          content:
            application/json:
              schema:
                title: DeleteBoardNotFoundResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                title: DeleteBoardErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                title: DeleteBoardServerErrorResponse
                $ref: './common.yaml#/components/schemas/ErrorResponse'
...